Make web socket host port optional.
Accept empty strings as a valid port in the GUI and derive the default
port for the given protocol at the connection level. This allows the
Host header of the HTTP/HTTPS connection to contain or omit the port
used depending on its presence. Useful when the web socket server has
capricious host validation logic